// src/app/shared/models/est_taksasi.model.ts
export class EstTaksasi {
	id?: number;

  blok_id?: number;
  afdeling_id?: number;
  lokasi_id?: number;
  tanggal?: string;

  ha_sisa?: number;
  ha_besok?: number;
  jumlah_pokok?: number;

  persen_buah_matang?: number;
  jjg_output?: number;

  hk?: number;
  bjr?: number;

  berat_kg?: number;
  seksi_panen?: number;
}